Moved ChangeBuffer to src/utils/.
